Guard against NPE in regex expection
If you expect a regex of a string but your test fails such that the actual value is nil, you get an exception from the Java regex pattern matcher which obscures the actual test failure. This Pull Request guards the regex test so that a nil value will fail the test without an exception, making the test result much more useful.