Fix #80 by allowing predicates on exceptions
This makes an expected predicate a special case for actual exceptions so
that you can expect more of an exception (than just its class).

The output from a predicate that fails on an exception is:

* The exception as the actual value (including stack trace)

* The message that "exception thrown by <actual expression> is not ..."

(showing the predicate that failed).

This seemed to be the best compromise after some experimentation.