2.2.0-rc2 substantial rework of underlying clojure.test support
Switched from using clojure.test/report to clojure.test/do-report for compatibility with Cursive.
Removed line/file calculations (since do-report handles that).
Ensured that :error reports are actually exceptions (and adjusted to :fail if not).

Still needs a fair bit of cleaning up.
Still not happy with the translate to expect-test piece (but wary of breaking code that uses both namespaces).