Note (potentially) breaking change in how namespaces should be used
Now that expect is part of expectations.clojure.test -- and it's _different_ to expectations/expect -- it is important not to refer in the latter. For convenience, all of the public symbols in expectations are now available directly in expectations.clojure.test so you only need to rely on one namespace.

However, if you are just referring defexpect from e.c.t then your code will break since expect is now a macro, not just a symbol that gets translated by defexpect.